refactor: Step 분리 (#11)
