+++ BBANGZIP/BBANGZIP/Sources/Presentation/ProgressBar/Step.swift
@@ -1,0 +1,26 @@
+//
+//  Step.swift
+//  BBANGZIP
+//
+//  Created by 최유빈 on 1/14/25.
+//  Copyright © 2025 com.bbangzip. All rights reserved.
+//
+
+import SwiftUI
+
+enum Step: Int{
+    case First = 1
+    case Second = 2
+    case Third = 3
+    
+    var percentage: CGFloat {
+        switch self {
+        case .First:
+            0.04
+        case .Second:
+            0.5
+        case .Third:
+            1
+        }
+    }
+}
